Feat: conserta códigos de respostas para casos de erro

+++ src/services/userService.ts
@@ -1,4 +1,6 @@
 import { UserDTO } from '../dtos/userDTO';
+import UserConflictError from '../errors/userConflictError';
+import userNotFoundError from '../errors/userNotFoundError';
 import { User } from '../models/user';
 import UserRepository from '../repositories/userRepository';
 
@@ -10,6 +12,10 @@ class UserService {
   }
 
   async createUser(userDTO : UserDTO ): Promise<User>  {
+    const existingUser = await this.userRepository.getUserByEmail(userDTO.email);
+    if(existingUser)
+      throw new UserConflictError();
+
     const user = new User(
       userDTO.name, 
       userDTO.email, 
@@ -20,7 +26,10 @@ class UserService {
   }
 
   async getUser(id: string): Promise<User>  {
-    return await this.userRepository.getUser(id);
+    const user = await this.userRepository.getUser(id);
+    if (!user)
+      throw new userNotFoundError();
+    return user;
   }
 
   async getAllUsers(): Promise<User[]>  {
@@ -28,12 +37,24 @@ class UserService {
   }
 
   async updateUser(id: string, name: string, email: string, password: string, admin: boolean): Promise<User> {
-    const userChanges = new User(name, email, password, admin);
+    const user = await this.userRepository.getUser(id);
+    if (!user)
+      throw new userNotFoundError();
 
+    if (user.email != email){
+      const users = await this.userRepository.getUserByEmail(email);
+      if (users) throw new UserConflictError();
+    }
+
+    const userChanges = new User(name, email, password, admin);
     return await this.userRepository.updateUser(id, userChanges);
   }
 
   async deleteUser(id: string): Promise<void>  {
+    const user = await this.userRepository.getUser(id);
+    if (!user)
+      throw new userNotFoundError();
+
     return await this.userRepository.deleteUser(id);
   }
 }
